requirements.txt
PyYAML==6.0
accelerate==0.21.0
datasets==2.12.0
deepspeed==0.10.0
einops==0.6.1
evaluate==0.4.0
peft @ git+https://github.com/huggingface/peft.git@0b62b4378b4ce9367932c73540349da9a41bdea8
protobuf==4.23.3
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
torch==2.0.1
fire==0.5.0
pillow==10.0.0
transformers @ git+https://github.com/huggingface/transformers@a6e6b1c622d8d08e2510a82cb6266d7b654f1cbf
isort
black
wandb
opencv-python